language-voyager/requirements.txt
# FastAPI and Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0  # Updated to include WebSocket support
websockets>=12.0  # Added explicit WebSocket support
python-multipart>=0.0.6
pydantic[email]>=2.4.2
email-validator>=2.1.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database and Cache
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
redis>=5.0.1
alembic>=1.12.1

# HTTP and Async
aiohttp>=3.8.5
yarl>=1.9.4
multidict>=6.0.4
async-timeout>=4.0.3

# GIS and Data Processing
shapely>=2.0.2
pyproj>=3.6.1
pandas>=2.0.0,<2.2.0
numpy>=1.21.6,<2.0.0
pillow>=9.5.0
pyarrow>=11.0.0
ujson>=3.0.0
cachetools>=5.3.0
geopy>=2.4.1

# Testing
pytest>=7.4.3
httpx>=0.25.1
pytest-asyncio>=0.21.1
